+++ sync-for-commerce/src/codatsynccommerce/sdkconfiguration.py
@@ -18,8 +18,8 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.1'
-    sdk_version: str = '0.31.0'
-    gen_version: str = '2.91.4'
+    sdk_version: str = '0.32.0'
+    gen_version: str = '2.101.0'
 
     def get_server_details(self) -> tuple[str, dict[str, str]]:
         if self.server_url: