+++ sync-for-commerce/RELEASES.md
@@ -529,3 +529,13 @@ Based on:
 - [python v0.31.0] sync-for-commerce
 ### Releases
 - [PyPI v0.31.0] https://pypi.org/project/codat-sync-for-commerce/0.31.0 - sync-for-commerce
+
+## 2023-09-07 00:04:01
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.79.0 (2.101.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.32.0] sync-for-commerce
+### Releases
+- [PyPI v0.32.0] https://pypi.org/project/codat-sync-for-commerce/0.32.0 - sync-for-commerce